refactor all utilities into a utilities folder
