+++ app.js
@@ -1,7 +1,7 @@
 import express from 'express'
 import config from './config.js'
 import cors from 'cors'
-import logger from './logger.js'
+import logger from './utilities/logger.js'
 
 import v1Initialize from './routes/v1/initialize.js'
 import v1Engines from './routes/v1/engines.js'
