+++ engines/quantitative/engine.js
@@ -1,6 +1,5 @@
-import { LLMWrapper } from '../../utils.js';
 import QuantitativeEngineBrain from './QuantitativeEngineBrain.js'
-import logger from '../../logger.js'
+import logger from '../../utilities/logger.js'
 
 class Engine {
     constructor() {
