+++ engines/seldon/SeldonBrain.js
@@ -1,4 +1,5 @@
-import utils, { LLMWrapper } from '../../utils.js'
+import utils from '../../utilities/utils.js'
+import { LLMWrapper } from '../../utilities/LLMWrapper.js'
 import { marked } from 'marked';
 
 class ResponseFormatError extends Error {
