+++ evals/categories/qualitativeCausalReasoning.js
@@ -16,7 +16,7 @@
  * @module categories/qualitativeCausalReasoning
  */
 
-import utils from '../../utils.js';
+import utils from '../../utilities/utils.js';
 
 /** generic prompt used for all tests */
 const prompt = "Please give me a causal loop diagram that captures the key variables and causal processes described in the background information.";
