+++ routes/v1/initialize.js
@@ -1,5 +1,5 @@
 import express from 'express'
-import utils from './../../utils.js'
+import utils from '../../utilities/utils.js'
 
 const router = express.Router()
